Got rid of validation.js
